+++ webapp/src/Menu/index.tsx
@@ -5,7 +5,7 @@ import { LanguageContext } from "../Language/locale";
 import "./index.scss";
 
 interface NavLink {
-    title: string | JSX.Element;
+    title: string;
     href: string;
 }
 
@@ -13,7 +13,7 @@ export default function Menu() {
     const locale = useContext(LanguageContext);
     const languages = { fr: "Français", en: "English" };
     const links: NavLink[] = [
-        { title: locale.translations("about_me"), href: "/" },
+        // { title: locale.translations("about_me"), href: "/" },
         { title: "Blog", href: "https://blog.senpaisilver.com" },
         { title: "Twitter", href: "https://twitter.com/SenpaiSilver" },
         { title: "YouTube", href: "https://youtube.com/@SenpaiSilver" },
