+++ webapp/src/Utility/Icon.scss
@@ -9,6 +9,7 @@
 }
 
 img.Icon {
-    height: 1em;
-    max-width: 1em;
+    height: 0.9em;
+
+    // &.icon-desktop, &.icon-server, &.icon-laptop { fill: white}
 }
